[Ref] serialization_utils: Silence signed-unsigned-comparison warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@15069 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ common/serialization_utils.cpp
@@ -258,7 +258,7 @@ void SsbWrite::WriteMapItem(const ID &id,
 void SsbWrite::IncrementWriteCounter()
 {
 	m_nCounter++;
-	if (m_nCounter >= (std::numeric_limits<uint16>::max() >> 2))
+	if(m_nCounter >= static_cast<uint16>(std::numeric_limits<uint16>::max() >> 2))
 	{
 		FinishWrite();
 		AddWriteNote(SNW_MAX_WRITE_COUNT_REACHED);
